# Improve build time for compact gate