--- a/ipa-step/src/hash.rs
+++ b/ipa-step/src/hash.rs
@@ -1,12 +1,10 @@
-use std::{
-    collections::BTreeMap,
-    hash::{DefaultHasher, Hash, Hasher},
-};
+use std::collections::BTreeMap;
 
 use proc_macro2::{Ident, TokenStream};
 use quote::{format_ident, quote, ToTokens};
+use syn::{parse_str, Path};
 
-use crate::CompactGateIndex;
+use crate::{CompactGateIndex, StepHasher};
 
 /// Builds a map of step strings to the corresponding compact gate index. Emits an array of tuples
 /// containing the hash and the index, sorted by hash. [`FromStr`] implementation for
@@ -16,34 +14,9 @@ use crate::CompactGateIndex;
 /// The complexity of this operation at compile time is O(n) and the cost is hash(str)*n.
 /// Runtime overhead is proportional to hash(str)+log(n).
 ///
-/// ## Hash collisions
-/// This currently uses the standard library [`Hasher`] interface to generate hashes that are limited
-/// to 8 bytes in size. This means that there is a small chance of hash collisions that increases
-/// with the overall increase in number of steps.
-///
-/// If that happens, a list of potential mitigations include
-/// * salt the hash (see code below),
-/// * use a different hash algorithm
-/// * use a perfect hash function.
-///
-/// The latter may result in more work to be done at runtime, but likely shouldn't impact
-/// the latency as long as [`FromStr`] is only used to create HTTP requests.
-///
-/// This implementation panics if it detects hash collisions.
-///
-/// [`FromStr`]: std::str::FromStr
 pub(crate) struct HashingSteps {
     inner: BTreeMap<u64, CompactGateIndex>,
     type_name: Ident,
-}
-
-fn hash(s: &str) -> u64 {
-    let mut hasher = DefaultHasher::default();
-    // If you get a collision, you can avoid that by adding:
-    // > hasher.write_u8(0);
-    // Tweak the value or remove the tweak if you get another collision.
-    hasher.write_str(s);
-    hasher.finish()
 }
 
 impl HashingSteps {
@@ -59,7 +32,7 @@ impl HashingSteps {
     /// if the step already added or if there is a hash collision with any of the steps already
     /// added.
     pub fn hash(&mut self, step: &str, gate: CompactGateIndex) {
-        let h = hash(step);
+        let h = step.hash_step();
         if let Some(old_val) = self.inner.insert(h, gate) {
             panic!("Hash collision for {step}: {h} => {old_val} and {gate}. Check that there are no duplicate steps defined in the protocol.");
         }
@@ -75,6 +48,12 @@ impl ToTokens for HashingSteps {
         let lookup_type = self.lookup_type();
         let sz = self.inner.len();
         let hashes = self.inner.iter().map(|(h, i)| quote! {(#h, #i)});
+        let hasher_path: Path = parse_str(
+            &std::any::type_name::<dyn StepHasher>()
+                .strip_prefix("dyn ")
+                .expect("typename for traits returns the path with dyn prefix"),
+        )
+        .unwrap();
 
         tokens.extend(quote! {
             #lookup_type {
@@ -88,15 +67,9 @@ impl ToTokens for HashingSteps {
 
             impl #lookup_type {
                 fn find(&self, input: &str) -> Option<u32> {
-                    let h = Self::hash(input);
+                    use #hasher_path;
+                    let h = input.hash_step();
                     self.inner.binary_search_by_key(&h, |(hash, _)| *hash).ok().map(|i| self.inner[i].1)
-                }
-
-                /// This must be kept in sync with proc-macro code that generates the hash.
-                fn hash(s: &str) -> u64 {
-                    let mut hasher = ::std::hash::DefaultHasher::default();
-                    ::std::hash::Hash::hash(s, &mut hasher);
-                    ::std::hash::Hasher::finish(&hasher)
                 }
             }
         });