I was investigating why build with compact gate takes that long and everything pointed to large match statements inside `FromStr` and `AsRef` trait implementations for `CompactGate`. The major bottleneck was `AsRef` match statement and overall the size of match statement impacts build time exponentially. Once the number of steps crosses the 70k boundary, release builds do not finish. 

The fix was to replace the match arms with something else for both `FromStr` and `AsRef`. Latter can be done with a simple array indexed by compact gate index, so it is very straightforward. That alone got the compile time down from ~4mins to 50s for 20k steps.

Fixing `FromStr` is more involved because it requires a lookup table. I've experimented with PHF and an array of tuples `(hash(str), gate index)`, sorted by hash value. Both have similar compilation times and I ended up with the array implementation to reduce the runtime cost a bit - PHF may require a few hash evaluations on a given string at runtime and array needs just one + binary search inside a sorted array which should be fast.

## Compile time impact

The following table shows the build times for `cargo bench --bench oneshot_ipa --no-default-features --features "enable-benches compact-gate"` command. For debug builds I used `cargo test --bench oneshot_ipa --no-default-features --features "enable-benches compact-gate"`

I measured the impact of the second build without the fix implemented in #1086 - that allowed me to capture the overhead of compact gate generation and compilation. 


| | debug build (fresh) | debug build (second) | release build (fresh) | release build (second) |
|-|---------------------|----------------------| --- | -- |
| before | 1m 18s  | 14.56s | 4m 02s | 3m 49s |
| after    | 1m 07s  | 10.18s | **1m 30s** | **23s** | 

As it can be seen, compiler optimizations cause the slowdown. Hinting `rustc` that there is nothing to optimize by using arrays improve optimized build times. 

## Runtime impact

I didn't anticipate any impact on `as_ref` implementation because that's just an array lookup and was more concerned about `FromStr` as it now requires a hash + binary search. Fortunately we only use it in HTTP deserialization and only when we receive a new `step` request, so the impact should be proportional to the number of gates we use and limited to one time overhead per gate creation.

We don't have reliable benchmarks to test it at scale, so I used our integration test to validate.

`cargo test --test helper_networks https_semi_honest_ipa --no-default-features --features "cli web-app real-world-infra test-fixture compact-gate"`

for release build

`cargo test --release --test helper_networks https_semi_honest_ipa --no-default-features --features "cli web-app real-world-infra test-fixture compact-gate"`

| | debug | release |
|-|---------------------|----------------------| 
| before | 6.63s  | 1.54s |
| after    | 4.65s |  1.52s |